// src/web/CSSPWebToolsAng/src/app/auth/register/register.interfaces.ts
export interface RegisterText
{
    Register: string;
    FirstName: string;
    FirstNameIsRequired: string;
    LastName: string;
    LastNameIsRequired: string;
    UserName: string;
    UserNameIsRequired: string;
    Password: string;
    PasswordIsRequired: string;
    PasswordMustBeAtLeast6Characters: string;
    Cancel: string;
    Locale: string;
}